refactor: All the build tools + docs
So I did a lot and was doing it all over the place all the times, and I
was very lax about commits. So I’m sorry.

I’ve added documentation for each of the scripts that we have. All of
them do the thing that we expect them to (as far as I know), and we now
have a nice npm run help thing too.